pyproject.toml
[tool.poetry]
name = "strato-spin"
version = "0.1.0"
description = "A YAML-based multi-cloud infrastructure deployment tool"
authors = ["William An <anqfsh@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "strato_spin"}]

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.26.0"
pyyaml = "^6.0"
poetry-core = "^1.6.0"
click = "^8.1.3"
azure-identity = "^1.12.0"
azure-mgmt-resource = "^22.0.0"
google-cloud-storage = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "^23.3.0"
isort = "^5.12.0"

[tool.poetry.scripts]
strato-spin = "strato_spin.cli:deploy"

[tool.poetry.urls]
"Repository" = "https://github.com/your-org/strato-spin"
